Create autoIGV1.3.py
Cleaned up some of the error messages a bit and fixed the handling of filenames with whitespace in them.  This will require IGV 2.2.38 or later in order to work properly due to a bug in previous versions' handling of whitespaces in file and pathnames.  This version also remembers bad bamfile paths and will avoid trying to load a bamfile during a run if it was previously unsuccessful (saves time and avoids repeated errors). Finally, I set this version to capture the first image in a set of individual images at a locus while taking the group image (since the first load generates the single).